pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "contractllm"
version = "1.0.0"
description = "Typed, versioned, regression-safe contracts for LLM calls"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "Qazi Tabish Firoz Ahmed", email = "tabishkazi561@gmail.com"},
]

# CORE dependencies — the absolute minimum to import the package.
# No provider SDKs here. Users install only what they need.
dependencies = [
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
openai = [
    "openai>=1.0.0",
]
anthropic = [
    "anthropic>=0.20.0",
]
cli = [
    "typer>=0.9.0",
    "rich>=13.0.0",
]
examples = [
    "python-dotenv>=1.0.0",
]
all = [
    "openai>=1.0.0",
    "anthropic>=0.20.0",
]
full = [
    "openai>=1.0.0",
    "anthropic>=0.20.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
]
dev = [
    "openai>=1.0.0",
    "anthropic>=0.20.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
contractllm = "contractllm.cli.main:app"

[project.urls]
Homepage = "https://github.com/tabishqazi/contractllm"
Repository = "https://github.com/tabishqazi/contractllm"
Issues = "https://github.com/tabishqazi/contractllm/issues"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
line-length = 88
target-version = "py311"
